[Update] 구조 업데이트에 따른 모듈 변화
