+++ presentation/src/main/java/com/dsm/dms/presentation/di/module/MainActivityModule.kt
@@ -2,11 +2,15 @@ package com.dsm.dms.presentation.di.module
 
 import com.dsm.dms.presentation.di.module.main.MainModule
 import com.dsm.dms.presentation.di.module.main.MainStaticModule
-import com.dsm.dms.presentation.di.module.sign.SignModule
-import com.dsm.dms.presentation.di.module.sign.SignStaticModule
+import com.dsm.dms.presentation.di.module.main.apply.extension.ApplyExtensionFloorDetailModule
+import com.dsm.dms.presentation.di.module.main.login.LoginStaticModule
+import com.dsm.dms.presentation.di.module.main.register.RegisterModule
+import com.dsm.dms.presentation.di.module.main.register.RegisterStaticModule
 import com.dsm.dms.presentation.di.scope.FragmentScope
 import com.dsm.dms.presentation.ui.fragment.MainFragment
-import com.dsm.dms.presentation.ui.fragment.sign.SignFragment
+import com.dsm.dms.presentation.ui.fragment.apply.extension.ApplyExtensionFloorDetailFragment
+import com.dsm.dms.presentation.ui.fragment.sign.login.LoginFragment
+import com.dsm.dms.presentation.ui.fragment.sign.register.RegisterFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
@@ -17,6 +21,14 @@ abstract class MainActivityModule {
     abstract fun mainFragment(): MainFragment
 
     @FragmentScope
-    @ContributesAndroidInjector(modules = [SignModule::class, SignStaticModule::class])
-    abstract fun signFragment(): SignFragment
+    @ContributesAndroidInjector(modules = [LoginStaticModule::class])
+    abstract fun loginFragment(): LoginFragment
+
+    @FragmentScope
+    @ContributesAndroidInjector(modules = [RegisterModule::class, RegisterStaticModule::class])
+    abstract fun registerFragment(): RegisterFragment
+
+    @FragmentScope
+    @ContributesAndroidInjector(modules = [ApplyExtensionFloorDetailModule::class])
+    abstract fun applyExtensionsFloorDetailFragment(): ApplyExtensionFloorDetailFragment
 }
