+++ presentation/src/main/java/com/dsm/dms/presentation/di/module/main/apply/ApplyModule.kt
@@ -1,10 +1,12 @@
 package com.dsm.dms.presentation.di.module.main.apply
 
-import com.dsm.dms.presentation.di.module.main.apply.extension.ApplyExtensionStudyFloorListModule
+import com.dsm.dms.presentation.di.module.main.apply.extension.ApplyExtensionFloorDetailModule
+import com.dsm.dms.presentation.di.module.main.apply.extension.ApplyExtensionFloorListModule
 import com.dsm.dms.presentation.di.module.main.apply.main.ApplyMainModule
 import com.dsm.dms.presentation.di.module.main.apply.staying.ApplyStayingModule
 import com.dsm.dms.presentation.di.scope.ApplyFragmentScope
-import com.dsm.dms.presentation.ui.fragment.apply.extension.ApplyExtensionStudyFloorListFragment
+import com.dsm.dms.presentation.ui.fragment.apply.extension.ApplyExtensionFloorDetailFragment
+import com.dsm.dms.presentation.ui.fragment.apply.extension.ApplyExtensionFloorListFragment
 import com.dsm.dms.presentation.ui.fragment.apply.main.ApplyMainFragment
 import com.dsm.dms.presentation.ui.fragment.apply.staying.ApplyStayingFragment
 import dagger.Module
@@ -21,6 +23,6 @@ abstract class ApplyModule {
     abstract fun applyMainFragment(): ApplyMainFragment
 
     @ApplyFragmentScope
-    @ContributesAndroidInjector(modules = [ApplyExtensionStudyFloorListModule::class])
-    abstract fun applyExtensionStudyFloorListFragment(): ApplyExtensionStudyFloorListFragment
+    @ContributesAndroidInjector(modules = [ApplyExtensionFloorListModule::class])
+    abstract fun applyExtensionFloorListFragment(): ApplyExtensionFloorListFragment
 }
