--- presentation/src/main/java/com/dsm/dms/presentation/di/scope/SignFragmentScope.kt
@@ -1,7 +1,0 @@
-package com.dsm.dms.presentation.di.scope
-
-import javax.inject.Scope
-
-@Scope
-@kotlin.annotation.Retention(AnnotationRetention.RUNTIME)
-annotation class SignFragmentScope
